+++ geokit/src/math/polynomial.rs
@@ -1,11 +1,11 @@
-use std::ops::{Add, Mul};
-
+use super::Float;
 use num::{Num, Zero};
+use std::ops::{Add, Mul};
 
 /// A polynomial of a single variable of degree less than `D`.
 /// All coefficients are stored on the stack.
 #[derive(Copy, Clone, Debug, PartialEq)]
-pub struct Polynomial<const D: usize, T = f64>([T; D]);
+pub struct Polynomial<const D: usize, T = Float>([T; D]);
 
 impl<const D: usize, T> Polynomial<D, T>
 where
@@ -35,8 +35,8 @@ where
     }
 }
 
-impl<const D: usize> Polynomial<D, f64> {
-    pub fn fast_eval_at(&self, x: f64) -> f64 {
+impl<const D: usize> Polynomial<D, Float> {
+    pub fn fast_eval_at(&self, x: Float) -> Float {
         self.0
             .into_iter()
             .rev()
@@ -50,28 +50,29 @@ mod tests {
 
     use crate::math::complex::Complex;
     use crate::math::polynomial::Polynomial;
+    use crate::math::Float;
 
     #[test]
     fn eval_at() {
         // eval_at(T)
         assert_eq!(Polynomial::new([0.0; 3]).eval_at(1.0), 0.0);
         assert_eq!(
-            Polynomial::<3, f64>::new([0.0; 3]).fast_eval_at(1.0f64),
+            Polynomial::<3, Float>::new([0.0; 3]).fast_eval_at(1.0f64),
             0.0
         );
         assert_eq!(Polynomial::new([1., -1.0, 1.0]).eval_at(1.0), 1.0);
         assert_eq!(
-            Polynomial::<3, f64>::new([1., -1.0, 1.0]).fast_eval_at(1.0f64),
+            Polynomial::<3, Float>::new([1., -1.0, 1.0]).fast_eval_at(1.0f64),
             1.0
         );
         assert_eq!(Polynomial::new([1.0, 0.0, 2.0]).eval_at(1.0), 3.0);
         assert_eq!(
-            Polynomial::<3, f64>::new([1.0, 0.0, 2.0]).fast_eval_at(1.0f64),
+            Polynomial::<3, Float>::new([1.0, 0.0, 2.0]).fast_eval_at(1.0f64),
             3.0
         );
         assert_eq!(Polynomial::new([1.0, 0.0, 2.0]).eval_at(2.0), 9.0);
         assert_eq!(
-            Polynomial::<3, f64>::new([1.0, 0.0, 2.0]).fast_eval_at(2.0f64),
+            Polynomial::<3, Float>::new([1.0, 0.0, 2.0]).fast_eval_at(2.0f64),
             9.0
         );
